chore: move docker-compose to research folder
